+++ docs/src/Avatar.doc.js
@@ -2,6 +2,8 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { Avatar, Box, Text } from 'gestalt';
+import keerthi from './avatars/keerthi.jpg';
+import shanice from './avatars/shanice.jpg';
 import PropTable from './components/PropTable';
 import Example from './components/Example';
 import PageHeader from './components/PageHeader';
@@ -53,7 +55,7 @@ type AvatarExProps = {
 };
 
 function AvatarEx(props: AvatarExProps) {
-  const name = 'Long';
+  const name = 'Keerthi';
   const { size, src } = props;
   return (
     <Box padding={1}>
@@ -72,8 +74,6 @@ AvatarEx.propTypes = {
   src: PropTypes.string,
 };
 
-const imageSrc = 'https://i.pinimg.com/avatars/long_1468294555_444.jpg';
-
 card(
   <Example
     description={`
@@ -83,8 +83,8 @@ card(
     defaultCode={`
 <Avatar
 size="md"
-src="${imageSrc}"
-name="Long"
+src="${keerthi}"
+name="Keerthi"
 />
 `}
     scope={{ Avatar }}
@@ -110,7 +110,7 @@ card(
   <Avatar name="Julia" />
 </Box>
 <Box column={4}>
-  <Avatar name="Long" src="${imageSrc}" />
+  <Avatar name="Keerthi" src="${keerthi}" />
 </Box>
 </Box>
   `}
@@ -127,16 +127,13 @@ card(
     name="Without an image"
     defaultCode={`
 <Avatar
-name="Long"
+name="Keerthi"
 size="lg"
 />
   `}
     scope={{ Avatar }}
   />
 );
-
-const joeyImage =
-  'https://i.pinimg.com/avatars/joeyzingarelli_1497376443_280.jpg';
 
 card(
   <Example
@@ -146,9 +143,9 @@ card(
     name="Verified"
     defaultCode={`
 <Avatar
-name="Joey"
+name="Shanice"
 size="lg"
-src="${joeyImage}"
+src="${shanice}"
 verified
 />
   `}