+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Patch
 
 * Internal: update `yarn.lock` (#152)
+* Docs: include images in repo (#151)
 
 </details>
 
@@ -19,7 +20,6 @@
 ### Minor
 
 * Button / SearchField / SegmentedControl / SelectList / Tabs / TextField: consistent sizing + improve Windows compatibility (#148)
-
 * Icon: Add new prop to Icon -- dangerouslySetSvgPath (#142)
 
 ## 0.67.0 (April 25, 2018)